Refactor sidebar view components
